requirements.txt
smolagents
pypdf
duckduckgo-search
markdownify
requests
python-dotenv
litellm
